Update racecar demo with modified asset loading and scaling for larger screens
Images can be loaded in ASTC from flash or a flash image file, as PNG files from the file system or from C arrays.
BT81x and BT82x are supported.
The screen can be scaled (using EVE scaling) depending on the panel size from the minimum 800x480 to 1920x1200.